backend/requirements.txt
annotated-types==0.7.0
blinker==1.9.0
CacheControl==0.14.1
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==44.0.0
distlib==0.3.9
filelock==3.16.1
firebase-admin==6.6.0
Flask==3.1.0
google-api-core==2.24.0
google-api-python-client==2.155.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-genai==0.2.1
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpcio==1.68.1
grpcio-status==1.68.1
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
msgpack==1.1.0
packaging==24.2
pillow==11.0.0
pipdeptree==2.23.4
pipenv==2024.4.0
platformdirs==4.3.6
proto-plus==1.25.0
protobuf==5.29.1
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.3
pydantic_core==2.27.1
PyJWT==2.10.1
pyparsing==3.2.0
python-dotenv==1.0.1
requests==2.32.3
rsa==4.9
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
virtualenv==20.27.1
websockets==14.1
Werkzeug==3.1.3
gunicorn==20.1.0
